Update user_account.sql
SMALLINT

+++ mariadb/user_account.sql
@@ -4,7 +4,7 @@ CREATE TABLE `user_account`
   `email` VARCHAR(255) NOT NULL,
   `first_name` VARCHAR(100) NOT NULL,
   `last_name` VARCHAR(100) NOT NULL,
-  `user_account_type_id` BIGINT UNSIGNED NOT NULL,
+  `user_account_type_id` SMALLINT UNSIGNED NOT NULL,
   `group_id` BIGINT UNSIGNED NOT NULL,
   `pbx_id` BIGINT UNSIGNED NOT NULL,
   `date_added` DATETIME NOT NULL,
